Refactor code for card click handler to work with new state, render and API fetch systems
